[BACKEND] : Minor fixes related to ROLES and Table Schema

+++ backend/src/main/java/com/buildgrid/MandiPro/repository/FirmRepository.java
@@ -7,6 +7,6 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface FirmRepository extends JpaRepository<Firm, Integer> {
+public interface FirmRepository extends JpaRepository<Firm, Long> {
     Optional<Firm> findByName(String name);
 }
